client/python/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatchling>=1.17",
]

[project]
name = "openlineage-python"
version = "1.15.0"
description = "OpenLineage Python Client"
readme = "README.md"
keywords = [
  "openlineage",
]
authors = [{ name = "OpenLineage", email = "info@openlineage.io" }]
requires-python = ">=3.8"
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "attrs>=20.0",
  "python-dateutil>=2.8.2",
  "pyyaml>=5.4",
  "requests>=2.20.0",
  "packaging>=21.0",
]
optional-dependencies.deserialize = [
  "cattrs>=23.2.3"
]
optional-dependencies.kafka = [
  "confluent-kafka>=2.1.1",
]
optional-dependencies.msk-iam = [
  "aws-msk-iam-sasl-signer-python>=1.0.1",
  "confluent-kafka>=2.1.1",
]
optional-dependencies.test = [
  "covdefaults>=2.3",
  "pytest>=7.3.1",
  "pytest-cov>=4.1",
  "pytest-mock>=3.10",
  "pyyaml>=5.4",
]

optional-dependencies.docs = [
  "sphinx>=7.0",
  "beautifulsoup4>=4.12.0",
  "sphinx-autodoc-typehints>=2.0.0",
  "pydata-sphinx-theme>=0.13"
]

[tool.hatch.build]
exclude = [".idea", ".ruff_cache", ".gitignore", "setup.cfg"]
targets.wheel = {packages = ["openlineage"]}

[tool.coverage]
html.show_contexts = true
html.skip_covered = false
report.fail_under = 91
run.parallel = true
run.plugins = ["covdefaults"]

[tool.mypy]
disable_error_code = "no-redef"
show_error_codes = true
overrides = [
  { ignore_missing_imports = true, module = [
  "confluent_kafka.*",
] } ]
strict = true

[tool.ruff]
line-length = 110
ignore = ["UP007"]
